talking_bi/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
google-generativeai==0.3.2
groq==1.1.2
mistralai==2.1.3
requests==2.33.0
